+++ db-manager/main.go
@@ -22,78 +22,6 @@ type Manifest struct {
 	Hash string `json:"hash"`
 }
 
-/***********************
-* This function inserts a record into the registry database
-* Input:
-*		*gorm.DB db
-* 		struct Record record
-* Returns:
-* 		None
-*
-************************/
-func insertRecord(db *gorm.DB, record *Manifest) (newRecord Manifest, Error error) {
-
-	result := db.Create(&record)
-
-	return *record, result.Error
-
-}
-
-/***********************
-* This function deletes a record by filename from the registry database.
-* Input:
-*		*gorm.DB db
-* 		string filename
-* Returns:
-* 		None
-*
-************************/
-func deleteRecord(db *gorm.DB, filename string) (Error error) {
-
-	result := db.Where("filename = ?", filename).Delete(&Manifest{})
-
-	if result.RowsAffected == 0 {
-		return result.Error
-	} else {
-		return nil
-	}
-
-}
-
-/***********************
-* This function queries the database for all the records. It returns an array of
-* all the records
-* Input:
-*		*gorm.DB db
-* Returns:
-* 		[]Record record
-*
-************************/
-func queryRecord(db *gorm.DB) (record []Manifest, Error error) {
-
-	result := db.Find(&record)
-
-	return record, result.Error
-
-}
-
-/***********************
-* This function queries the database for a file and returns the record for the file.
-* Input:
-*		*gorm.DB db
-* 		filename
-* Returns:
-* 		Record record
-*
-************************/
-func querySingleRecord(db *gorm.DB, filename string) (record Manifest, Error error) {
-
-	result := db.Where("filename = ?", filename).First(&record)
-
-	return record, result.Error
-
-}
-
 // Server Instance Configuration:
 // TODO: read this from a config.json or similar
 const index = 0 // the index of this DB instance
@@ -102,9 +30,6 @@ const VERSION = "v1"
 const PG_PASSWORD = "password"
 
 func main() {
-
-	http.HandleFunc("/api/"+VERSION+"/records/", recordHandler)
-	http.HandleFunc("/api/"+VERSION+"/records", recordsHandler)
 
 	// Connection to the database
 	dsn := fmt.Sprintf("host=localhost user=postgres password=%s dbname=registry%d port=5432 sslmode=disable TimeZone=UTC", PG_PASSWORD, index)
@@ -117,6 +42,9 @@ func main() {
 	if err != nil {
 		log.Fatal(err)
 	}
+
+	http.HandleFunc("/api/"+VERSION+"/records/", createRecordHandler(db))
+	http.HandleFunc("/api/"+VERSION+"/records", createRecordsHandler(db))
 
 	fmt.Printf("Server starting on port %s...\n", PORT)
 	err = http.ListenAndServe(net.JoinHostPort("localhost", PORT), nil)
